requirements.txt
# Usage: pip install -r requirements.txt
annotated-types==0.6.0
matplotlib==3.7.5
numpy==1.24.4
opencv-python==4.9.0.80
pandas==2.0.3
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.0    
pydantic-core==2.18.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.3.1
pyzed==4.1
requests==2.31.0
scipy==1.10.1
six==1.14.0
thop==0.1.1.post2209072238
torch==2.1.0a0+41361538.nv23.6
torchvision==0.16.1
tqdm==4.66.2
typing-extensions==4.11.0
tzdata==2024.1 
ultralytics==8.2.2
setuptools==45.2.0
